renamed router to RoshiRouter
